refactor evaluator
